+++ paimon-common/src/main/java/org/apache/paimon/fileindex/FileIndexer.java
@@ -24,10 +24,6 @@ import org.apache.paimon.types.DataType;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import java.util.HashMap;
-import java.util.Map;
-import java.util.ServiceLoader;
-
 /** File index interface. To build a file index. */
 public interface FileIndexer {
 
@@ -38,36 +34,7 @@ public interface FileIndexer {
     FileIndexReader createReader(byte[] serializedBytes);
 
     static FileIndexer create(String type, DataType dataType, Options options) {
-
-        FileIndexerFactory fileIndexerFactory = FileIndexerLoadFactory.load(type);
+        FileIndexerFactory fileIndexerFactory = FileIndexerFactoryUtils.load(type);
         return fileIndexerFactory.create(dataType, options);
-    }
-
-    /** Load factory to load FileIndexerFactory. */
-    class FileIndexerLoadFactory {
-
-        private static final Map<String, FileIndexerFactory> factories = new HashMap<>();
-
-        static {
-            ServiceLoader<FileIndexerFactory> serviceLoader =
-                    ServiceLoader.load(FileIndexerFactory.class);
-
-            for (FileIndexerFactory indexerFactory : serviceLoader) {
-                if (factories.put(indexerFactory.identifier(), indexerFactory) != null) {
-                    LOG.warn(
-                            "Found multiple FileIndexer for type: "
-                                    + indexerFactory.identifier()
-                                    + ", choose one of them");
-                }
-            }
-        }
-
-        static FileIndexerFactory load(String type) {
-            FileIndexerFactory fileIndexerFactory = factories.get(type);
-            if (fileIndexerFactory == null) {
-                throw new RuntimeException("Can't find file index for type: " + type);
-            }
-            return fileIndexerFactory;
-        }
     }
 }
